have test_tx check for actual payload
Without matching exact input with output, we rely on
tshark's dissector to ensure validity for us. If the UDP
payload can't be matched we know something went wrong. In
this case it looks like the 8787 firmware will mangle the
FC on transmission so tshark doesn't consider it a QoS
frame and chokes.